requirements.txt
django==2.1.7
pillow==5.4.1
psycopg2-binary==2.7.6.1
pytz==2018.9
django-simple-captcha